updates to kvm create script

+++ vps_kvm_create.sh
@@ -104,7 +104,9 @@ else
   sed s#"<features>"#"<features>\n    <hap/>"#g -i ${name}.xml
  fi
  rm -f ${name}.xml.backup
- sed s#"<source bridge='br0'\/>"#"<model type='virtio'/>\n          <source bridge='br0'/>"#g -i ${name}.xml
+ if [ ! -e /etc/apt ] || [ "$template" != "windows2" ]; then
+  sed s#"<source bridge='br0'\/>"#"<model type='virtio'/>\n          <source bridge='br0'/>"#g -i ${name}.xml
+ fi;
  /usr/bin/virsh define ${name}.xml
  if [ "$template" = "windows1" ]; then
   template=windows2
